+++ apps/api/src/main.ts
@@ -3,10 +3,14 @@ import { Logger, LoggerErrorInterceptor } from 'nestjs-pino';
 import { ValidationPipe, VersioningType } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
+import cookieParser from 'cookie-parser';
 import { AppModule } from './app.module';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule, { bufferLogs: true });
+
+  // Parse HTTP cookies
+  app.use(cookieParser());
 
   // Use nestjs-pino logger
   app.useLogger(app.get(Logger));